pyproject.toml
[project]
name = "what-are-we-doing"
version = "0.1.0"
description = "Shared versioned workspace with SLM oracle for multi-agent collaboration"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "watchdog>=4.0.0",
    "mcp>=1.0.0",
    "aiosqlite>=0.20.0",
    "zstandard>=0.22.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-tmp-files>=0.0.2",
]

[project.scripts]
wawd = "wawd.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["wawd"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
